Crate per_cpu

Contains PerCpuData, the data stored on a per-CPU basis in Theseus.

Each CPU has its own instance of PerCpuData, and each CPU’s instance can only be accessed by itself.

This per_cpu crate vs. the cpu_local_preemption crate

These two crates exist to solve a circular dependency problem: the crate that defines the per-CPU data structure (this per_cpu crate) must depend on all the foreign crates that define the types used for each field in the per-CPU data structure. However, those foreign crates also want to access these per-CPU states, which would require depending on this per_cpu crate. This would create a cyclic dependency, so we break it into two crates.

This crate per_cpu directly depends on many other kernel crates, specifically the ones that define the types needed for each field of PerCpuData.

If you want to add another piece of per-CPU data, you can do that here by modifying the PerCpuData struct, and then updating the const definitions of offsets and other metadata in cpu_local_preemption::FixedCpuLocal.

To actually access per-CPU data, do not use this crate, use cpu_local_preemption::CpuLocal instead.

The cpu_local_preemption crate is the “top-level” crate that is depended upon by each of the crates that needs to access per-CPU data.

cpu_local_preemption is a mostly standalone crate that does not depend on any of the specific types from other Theseus crates, which allows other Theseus crates to depend upon it.

cpu_local_preemption effectively decouples the definitions

This per_cpu crate also depends on cpu_local_preemption in order to initialize itself for each CPU right after that CPU has booted.

Structs

CpuLocalCpuId

An immutable type wrapper for this CPU’s unique ID, kept in CPU-local storage.

PerCpuData

The data stored on a per-CPU basis in Theseus.

Functions

init

Initializes the current CPU’s PerCpuData.